@@ -2,6 +2,7 @@ package eu.siacs.conversations.ui;
 
 import android.preference.CheckBoxPreference;
 import android.support.annotation.NonNull;
+import android.support.v4.content.ContextCompat;
 import android.support.v7.app.AlertDialog;
 import android.app.FragmentManager;
 import android.content.DialogInterface;
@@ -406,7 +407,7 @@ public class SettingsActivity extends XmppActivity implements
 	}
 
 	private void startExport() {
-		startService(new Intent(getApplicationContext(), ExportLogsService.class));
+		ContextCompat.startForegroundService(this, new Intent(this, ExportLogsService.class));
 	}
 
 	private void displayToast(final String msg) {
